Refactor input error handling and update component imports for improved structure and consistency
