+++ README.md
@@ -1,3 +1,4 @@
+# Laravel Boilerplate
 
 ## Getting Started
