+++ resources/js/components/app-logo.tsx
@@ -1,7 +1,5 @@
 import AppLogoIcon from '@/components/app-logo-icon';
 
-
-
 export default function AppLogo() {
     return (
         <>
