+++ resources/js/components/stocks/input-error.tsx
