+++ resources/js/layouts/auth/auth-split-layout.tsx
@@ -13,7 +13,7 @@ export default function AuthSplitLayout({ children, title, description }: PropsW
             <div className="flex flex-col gap-4 p-6 md:p-10">
                 <div className="flex justify-center gap-2 md:justify-start">
                     <Link href="/" className="flex items-center gap-2 font-medium">
-                        <div className="bg-primary text-primary-foreground flex size-8 items-center justify-center rounded-lg aspect-square">
+                        <div className="bg-primary text-primary-foreground flex aspect-square size-8 items-center justify-center rounded-lg">
                             <AppLogoIcon className="dark:text-primary-foreground size-4" />
                         </div>
                         Laravel Boilerplate
